feat: move dataclasses to k8s.models
 - Cluster
 - APIObjectInCluster
